Multi exposure report metadata (#4889)
Multi exposure report urls metadata
